+++ lib/impl/camel_caser/any.ex
@@ -1,5 +1,3 @@
 defimpl Digger.CamelCaser, for: Any do
-  def camel_case(any, _first_letter) do
-    any
-  end
+  def camel_case(any, _), do: any
 end