+++ lib/impl/stringifier/atom.ex
@@ -1,8 +1,8 @@
 defimpl Digger.Stringifier, for: Atom do
-  def stringify(atom, type: "key", stringify_key: true, stringify_value: _),
+  def stringify(atom, type: :key, stringify_key: true, stringify_value: _),
     do: atom |> Atom.to_string()
 
-  def stringify(atom, type: "value", stringify_value: true, stringify_value: _),
+  def stringify(atom, type: :value, stringify_value: true, stringify_value: _),
     do: atom |> Atom.to_string()
 
   def stringify(atom, _), do: atom